seppareted classes for mapping
